+++ packages/components/src/List/ListComposition.stories.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { action } from '@storybook/addon-actions';
 
 import { simpleCollection } from './ListComposition/collection';
@@ -691,6 +691,45 @@ export const SelectableItems = () => {
 						isSelected={isSelected}
 						onToggleAll={onToggleAll}
 						selectionToggle={(_, group) => onToggleItem(group)}
+					/>
+				</List.Manager>
+			</section>
+		</div>
+	);
+};
+
+export const SelectableItemsWithDisabledItems = () => {
+	const { isSelected, onToggleAll, onToggleItem } = List.hooks.useCollectionSelection(
+		simpleCollection,
+		[],
+		'id',
+	);
+
+	const getRowState = item => (item.id === 1 ? { disabled: true } : {});
+	const isToggleAllDisabled = items => {
+		for (let i = 0; i < items.length; i++) {
+			if (getRowState(items[i]).disabled) {
+				return true;
+			}
+		}
+		return false;
+	};
+	return (
+		<div className="virtualized-list">
+			<h1>List with selection disabled items</h1>
+			<p>
+				You can pass <b>getRowState</b> for disable a certain row. <br />
+				<b>isToggleAllDisabled</b> is used for disable "Select all" checkbox
+			</p>
+
+			<section style={{ height: '50vh' }}>
+				<List.Manager id="my-list" collection={simpleCollection}>
+					<CustomList
+						isSelected={isSelected}
+						onToggleAll={onToggleAll}
+						selectionToggle={(_, group) => onToggleItem(group)}
+						getRowState={getRowState}
+						isToggleAllDisabled={isToggleAllDisabled}
 					/>
 				</List.Manager>
 			</section>